Убран лишний импорт, добавлены проверки
Убрано ObjectsDoesNotExist, дописааа проверка на существование ученика и предмета

+++ scripts.py
@@ -1,7 +1,6 @@
 import random
 
 from datacenter.models import Schoolkid, Mark, Lesson, Chastisement, Commendation
-from django.core.exceptions import MultipleObjectsReturned, ObjectDoesNotExist
 
 COMMENDATIONS = [
     "Молодец!",
@@ -37,44 +36,46 @@ COMMENDATIONS = [
 ]
 
 
-
 def get_schoolkid(schoolkid_name):
     try:
         schoolkid = Schoolkid.objects.get(full_name__contains=schoolkid_name)
-    except Schoolkid.ObjectDoesNotExist:
-        print(f"Не найдено ученика с именем {schoolkid_name}")
-        return None
     except Schoolkid.MultipleObjectsReturned:
         print(f"Найдено больше одного ученика с именем {schoolkid_name}")
-        exit()
+        return None
+    except Schoolkid.DoesNotExist:
+        print(f"Не найдено ученика с именем {schoolkid_name}")
+        return None
     return schoolkid
 
 
 def fix_marks(schoolkid_name):
     schoolkid = get_schoolkid(schoolkid_name)
-    Mark.objects.filter(schoolkid=schoolkid, points__in=[2, 3]).update(points=5)
+    if not (schoolkid is None):
+        Mark.objects.filter(schoolkid=schoolkid, points__in=[2, 3]).update(points=5)
 
 
 def delete_chastisements(schoolkid_name):
     schoolkid = get_schoolkid(schoolkid_name)
-    Chastisement.objects.filter(schoolkid=schoolkid).delete()
+    if not (schoolkid is None):
+        Chastisement.objects.filter(schoolkid=schoolkid).delete()
 
 
 def create_commendation(schoolkid_name, subject_name):
     schoolkid = get_schoolkid(schoolkid_name)
-    try:
+    if not (schoolkid is None):
         last_lesson = Lesson.objects.filter(
             year_of_study=schoolkid.year_of_study,
             group_letter=schoolkid.group_letter,
             subject__title__contains=subject_name,
         ).order_by('-date').first()
-        commendation_text = random.choice(COMMENDATIONS)
-        Commendation.objects.create(
-            text=commendation_text,
-            schoolkid=schoolkid,
-            subject=last_lesson.subject,
-            teacher=last_lesson.teacher,
-            created=last_lesson.date,
-        )
-    except Lesson.ObjectDoesNotExist:
-        print (f"Нельзя создать похвалу. У ученика {schoolkid_name} нет уроков {subject_name}")
+        if not (last_lesson is None):
+            commendation_text = random.choice(COMMENDATIONS)
+            Commendation.objects.create(
+                text=commendation_text,
+                schoolkid=schoolkid,
+                subject=last_lesson.subject,
+                teacher=last_lesson.teacher,
+                created=last_lesson.date,
+            )
+        else:
+            print(f"Нельзя создать похвалу. У ученика {schoolkid_name} нет уроков {subject_name}")
